requirements.txt
asgiref==3.9.1
black==25.1.0
boto3==1.40.16
botocore==1.40.16
click==8.2.1
Django==5.2.5
django-cors-headers==4.7.0
django-extension==0.1.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
drf-yasg==1.21.10
inflection==0.5.1
jmespath==1.0.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
PyJWT==2.10.1
pymongo==4.14.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
coverage==7.4.1
ruff==0.12.10
s3transfer==0.13.1
shelve_cache==0.0.4
six==1.17.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
